% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.stats-data.R
\docType{data}
\name{cv.stats}
\alias{cv.stats}
\title{cross-validated MOSAIC runs}
\format{
A list of 13 lists as follows.
\describe{
\item{acc_DHS}{}
\item{acc_p300}{}
\item{acc_cgi}{}
\item{acc_CTCF}{}
\item{acc_promoter}{}
\item{acc_genebody}{}
\item{met_DHS}{}
\item{met_p300}{}
\item{met_cgi}{}
\item{met_CTCF}{}
\item{met_promoter}{}
\item{met_genebody}{}
\item{rna.mat}{}

}
}
\usage{
data(cv.stats)
}
\description{
A list of lists output of MOSAIC run for 50 rounds of 5-fold cross vaidation for
each of the 13 data types for k= 2 to 7. Sumamrized over Adjusted Mutual Information
(AMI) and Standardized Pooled Within Sum of Squares (SPWSS)
}
\keyword{datasets}
