% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.all-data.R
\docType{data}
\name{lineage.all}
\alias{lineage.all}
\title{lineage labels}
\format{
A character vector of lineage labels.
\describe{
}
}
\usage{
data(lineage.all)
}
\description{
A character vector of lineage classification
}
\keyword{datasets}
