% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage-data.R
\docType{data}
\name{stage}
\alias{stage}
\title{stage data used}
\format{
A character vector of stage labels.
\describe{
}
}
\usage{
data(stage)
}
\description{
A character vector of stage labels used as outcome in MOSAIC.
Visceral endoderms were removed.
}
\keyword{datasets}
